modularized product into own file
